\name{Leukemia_miR}
\alias{Leukemia_miR}
\docType{data}
\title{microRNA expression of acute leukemia phenotypes}
\description{
Data of human microRNA (miR) expression of 847 miRs from 80 acute myeloid (AML) and acute lymphoblastic (ALL) leukemia cell lines, 60 primary (patient) samples, and 50 normal control samples sorted by cell type (CD34+ HSPC, Granulocytes, Monocytes, T-cells and B-cells). Acute lymphoblastic leukemia samples are further classified by B-cell (B-ALL) and T-cell (T-ALL) subphenotypes.
}

\usage{data(Leukemia_miR)}

\details{
Two dataset versions are provided: the full dataset \code{Leuk_miR_full} (190 samples x 847 miRs) and the filtered dataset \code{Leuk_miR_filt} (140 samples x 370 miRs). Data available at GEO under Accession Number GSE51908.
}

\references{
Tan YS et al.
\emph{Regulation of RAB5C is important for the growth inhibitory effects of MiR-509 in human precursor-B acute lymphoblastic leukemia}, PLoS One (2014) 9:e111777.

Candia J et al.
  \emph{Uncovering low-dimensional, miR-based signatures of acute myeloid and lymphoblastic leukemias with a machine-learning-driven network approach},
  Converg Sci Phys Oncol (2015) 1:025002.\cr
}

\keyword{datasets}

