\name{QuickStartEx}
\alias{QuickStartEx}
\docType{data}
\title{synthetic dataset}
\description{
75 instances with 20 predictors and a numerical response to be used as a quick start example. 
}
\usage{data(QuickStartEx)}

\format{
A numerical matrix of predictors is provided with instances as rows and predictors as columns. A numerical response is provided 
as a quick start example for linear regression models; it can be easily discretized 
to serve as example for binary and multinomial models as well.
}

\keyword{datasets}

