\name{eNetXplorer-package}
\alias{eNetXplorer-package}
\alias{eNetXplorer}
\docType{package}
\title{Exploration of elastic net families for generalized linear models
%\packageTitle{eNetXplorer}
}
\description{
%This package provides a quantitative toolkit to explore elastic net families and to uncover correlates contributing to prediction under a cross-validation framework. Parameter importance is evaluated by flexible criteria based on out-of-bag predictions, which are assessed via user-defined quality functions. Statistical significance is assigned to each model by comparison to a set of null models generated by random permutations of the sample labels; analogous permutation procedures were used to assess significance for the contribution of individual parameters to prediction. The package provides a set of standard plots, summary statistics and output tables. It fits linear, binomial (logistic) and multinomial models.
Provides a quantitative toolkit to explore elastic net families and to uncover correlates contributing to prediction under a cross-validation framework. Fits linear, binomial (logistic) and multinomial models.
%\packageDescription{eNetXplorer}
}
%\details{
\details{
\tabular{ll}{
Package: \tab eNetXplorer\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2018-04-05\cr
License: \tab GPL-3\cr
}
Functions:

\code{eNetXplorer}\cr
\code{plot.eNetXplorer}\cr
\code{pdf.summary}\cr
\code{eNetXplorerExport}\cr

}
%The DESCRIPTION file:
%\packageDESCRIPTION{eNetXplorer}
%\packageIndices{eNetXplorer}
%~~ An overview of how to use the package, including the most important functions ~~
%}
\author{Julian Candia and John S. Tsang\cr 
Maintainer: Julian Candia \email{julian.candia@nih.gov}}

%\author{
%\packageAuthor{eNetXplorer}
%
%Maintainer: \packageMaintainer{eNetXplorer}
%}
\references{
%~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
\dontrun{
data(QuickStartExample)
result = eNetXplorer(x=QuickStartExample$predictor,y=QuickStartExample$response,family="gaussian",n_run=25,n_perm_null=20,seed=111)
plot(result,plot.type="measured.vs.oob",alpha=4)
suppressWarnings(pdf.summary(result))
eNetXplorerExport(result)
}
}
